requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
tensorflow==2.13.0
scipy==1.11.4
Pillow==10.0.1
scikit-learn==1.3.2
gunicorn==21.2.0